working commit, before stretch goals

+++ src/com/skilldistillery/jets/entities/Passenger.java
@@ -2,7 +2,7 @@ package com.skilldistillery.jets.entities;
 
 public class Passenger extends Jet{
 
-	public Passenger(String type, String model, double speed, double price, int range) {
+	public Passenger(String type, String model, double speed, long price, int range) {
 		super(type, model, speed, price, range);
 		
 	}
